Add time logic to refactored version
