+++ lib/printer.rb
@@ -1,20 +1,23 @@
 require './lib/player_guess_matcher'
-# require './lib/time'
+
 
 class Printer
-  attr_reader :winning_code, :time_taken_to_win, :elapsed_total
+  attr_reader :winning_code, :time_taken_to_win
   attr_accessor :guess_count, :player_guess
 
-  def initialize(winning_code, guess_count, player_guess) #time_taken_to_win) # change nil if need be
+  def initialize(winning_code)
     @winning_code = winning_code
-    @guess_count = guess_count
-    @player_guess = player_guess
-    # @time_taken_to_win = time_taken_to_win
-    @instance_of_matcher = PlayerGuessMatcher.new(@winning_code, @player_guess)
-    # @instance_of_time = Time.new(@starting_time, @end_time)
-    # @instance_of_mastermind = Mastermind.new(@guess_count, @winning_code)
-    @elapsed_total = elapsed_total
+    @instance_of_matcher = PlayerGuessMatcher.new(@winning_code)
+  end
 
+  def player_guess_error_message(player_guess)
+      player_guess.chars.each do |color|
+        if !(color).match(/[rybg]/)
+          puts "------------------"
+          puts "ERROR: Choose four colors: R G B Y"
+          puts "Guess again:"
+        end
+    end
   end
 
   def welcome_message
@@ -28,9 +31,6 @@ class Printer
     @guesses_or_guess = guess_count > 1 ? "guesses" : "guess"
   end
 
-  # def handle_minute_singular_or_plural(end_time_minutes, start_time_minutes)
-  #   @minutes_or_minute = end_time_minutes - start_time_minutes = 1 ? "minute" : "minutes" #instance variables?
-  # end
 
   def guess_feedback(player_guess, guess_count) #had to add ability to pass in argument into method - because player_guess is user input??
     puts "------------------"
@@ -48,9 +48,6 @@ class Printer
 
     handle_guess_singular_or_plural(guess_count)
     # handle_minute_singular_or_plural(end_time_minutes, start_time_minutes)
-    # minutes_elapsed = end_time_minutes - start_time_minutes
-    # seconds_elapsed = end_time_seconds - start_time_seconds
-    # binding.pry
     win_message = "Congratulations! You guessed the sequence '#{@winning_code}' in #{guess_count} #{@guesses_or_guess} over #{time_message}!
     Do you want to (p)lay again or (q)uit?"
